requirements.txt
fastapi>=0.104.0
jinja2>=3.1.0
lighthive>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0
uvicorn>=0.24.0
pyyaml>=6.0.1
pydantic>=2.5.0
bcrypt>=4.1.0
python-multipart>=0.0.9
